Refactor: Remove Limelight class and related logic
The `Limelight.java` class has been deleted.

All related code for Limelight integration, including auto-centering logic, has been removed from `TeleOpMain.java` and `MecanumChassis.java`. This simplifies the TeleOp and Chassis classes by removing vision-based functionality.
