+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Resources/MecanumChassis.java
@@ -481,226 +481,11 @@ public class MecanumChassis {
     }
 
     /**
-     * Center the robot on a target using Limelight data and encoder-based movement
-     * This method rotates the robot to face the target based on Limelight's tx value
-     * 
-     * @param limelight The Limelight instance to get targeting data from
-     * @param rotationPower The power to use for rotation (0.0 to 1.0)
-     * @param timeoutSeconds Maximum time to spend centering (0 = no timeout)
-     * @return true if successfully centered, false if target lost or timeout
-     */
-    public boolean centerOnLimelightTarget(org.firstinspires.ftc.teamcode.Game.Limelight limelight,
-                                           double rotationPower, double timeoutSeconds) {
-        if (limelight == null) {
-            opmode.telemetry.addData("Error", "Limelight is null");
-            opmode.telemetry.update();
-            return false;
-        }
-
-        // Update limelight to get latest data
-        limelight.update();
-        
-        // Check if target is valid
-        if (!limelight.hasValidTarget()) {
-            opmode.telemetry.addData("Limelight Center", "No valid target detected");
-            opmode.telemetry.update();
-            return false;
-        }
-
-        // Use raw angle (bypasses threshold) to always attempt centering when button is pressed
-        double rotationAngle = limelight.getCenteringRotationAngleRaw();
-        
-        opmode.telemetry.addData("Limelight Center", "Rotation Angle: %.2f degrees", rotationAngle);
-        opmode.telemetry.update();
-
-        // If angle is very small (less than 0.2 degrees), consider it centered
-        if (Math.abs(rotationAngle) < 0.2) {
-            opmode.telemetry.addData("Limelight Center", "Already centered (< 0.2 deg)");
-            opmode.telemetry.update();
-            return true;
-        }
-
-        opmode.telemetry.addData("Limelight Center", "Rotating by: %.2f degrees", rotationAngle);
-        opmode.telemetry.update();
-
-        // Use the existing rotate method to rotate by the calculated angle
-        // Note: rotate() uses IMU, so we're rotating relative to current heading
-        // rotate() already sets motors to RUN_WITHOUT_ENCODER, so manual control is preserved
-        rotate(rotationAngle, rotationPower);
-
-        // Verify we're now centered
-        limelight.update();
-        if (limelight.isCentered()) {
-            opmode.telemetry.addData("Limelight Center", "Successfully centered!");
-            opmode.telemetry.update();
-            return true;
-        }
-
-        opmode.telemetry.addData("Limelight Center", "Centering may need adjustment");
-        opmode.telemetry.update();
-        return true; // Return true even if not perfectly centered, as we did our best
-    }
-
-    /**
-     * Center the robot on a Limelight target with default power
-     * @param limelight The Limelight instance
-     * @return true if successfully centered
-     */
-    public boolean centerOnLimelightTarget(org.firstinspires.ftc.teamcode.Game.Limelight limelight) {
-        return centerOnLimelightTarget(limelight, 0.5, 0);
-    }
-
-    /**
-     * Continuously center on a Limelight target until centered or timeout
-     * This will keep adjusting until the robot is within the centering threshold
-     * 
-     * @param limelight The Limelight instance
-     * @param rotationPower The power to use for rotation
-     * @param maxIterations Maximum number of centering attempts (0 = no limit, but check opModeIsActive)
-     * @return true if successfully centered
-     */
-    public boolean centerOnLimelightTargetContinuous(org.firstinspires.ftc.teamcode.Game.Limelight limelight,
-                                                      double rotationPower, int maxIterations) {
-        if (limelight == null) {
-            return false;
-        }
-
-        int iterations = 0;
-        double minAngleThreshold = 0.5; // Minimum angle to bother rotating
-
-        while (opmode.opModeIsActive() && (maxIterations == 0 || iterations < maxIterations)) {
-            limelight.update();
-            
-            if (!limelight.hasValidTarget()) {
-                opmode.telemetry.addData("Limelight Center", "Target lost");
-                opmode.telemetry.update();
-                return false;
-            }
-
-            double rotationAngle = limelight.getCenteringRotationAngle();
-            
-            // If centered, we're done
-            if (Math.abs(rotationAngle) < minAngleThreshold || limelight.isCentered()) {
-                opmode.telemetry.addData("Limelight Center", "Centered! Iterations: %d", iterations);
-                opmode.telemetry.update();
-                return true;
-            }
-
-            // Rotate by the calculated angle
-            rotate(rotationAngle, rotationPower);
-            iterations++;
-
-            opmode.telemetry.addData("Limelight Center", "Iteration: %d, Angle: %.2f", iterations, rotationAngle);
-            opmode.telemetry.update();
-            
-            opmode.sleep(100); // Small delay between iterations
-        }
-
-        return limelight.isCentered();
-    }
-
-    /**
      * Restore motors to manual control mode (RUN_WITHOUT_ENCODER)
      * This should be called after encoder-based movements to allow manual driving
      */
     public void restoreManualControl() {
         run_without_encoders_all();
-    }
-
-    /**
-     * Center the robot's position using Limelight by strafing left/right (encoder-based)
-     * This calculates the distance needed to strafe based on tx angle and distance to target
-     * 
-     * @param limelight The Limelight instance
-     * @param movePower Power to use for movement (0.0 to 1.0)
-     * @param distanceToTarget Estimated distance to target in inches (used to calculate strafe distance)
-     * @return true if successfully centered
-     */
-    public boolean centerOnLimelightTargetByStrafing(org.firstinspires.ftc.teamcode.Game.Limelight limelight,
-                                                      double movePower, double distanceToTarget) {
-        if (limelight == null) {
-            opmode.telemetry.addData("Error", "Limelight is null");
-            opmode.telemetry.update();
-            return false;
-        }
-
-        limelight.update();
-        
-        if (!limelight.hasValidTarget()) {
-            opmode.telemetry.addData("Limelight Center", "No valid target detected");
-            opmode.telemetry.update();
-            return false;
-        }
-
-        // Get the horizontal offset angle directly using getTx() method
-        double tx = limelight.getTx();
-        opmode.telemetry.addData("Limelight Center", "Tx: %.2f degrees", tx);
-        opmode.telemetry.update();
-        
-        // If already well-centered (within 0.5 degrees), return success
-        if (Math.abs(tx) < 0.5) {
-            opmode.telemetry.addData("Limelight Center", "Already well-centered (< 0.5 deg)");
-            opmode.telemetry.update();
-            restoreManualControl(); // Restore manual control mode
-            return true;
-        }
-
-        // Convert angle to strafe distance using trigonometry
-        // distance = tan(angle) * distanceToTarget
-        // Note: tx is in degrees, so convert to radians
-        double angleRadians = Math.toRadians(tx);
-        double strafeDistanceInches = Math.tan(angleRadians) * distanceToTarget;
-
-        opmode.telemetry.addData("Limelight Center", "Tx: %.2f deg, Strafe: %.2f inches", tx, strafeDistanceInches);
-        opmode.telemetry.update();
-
-        // Determine direction and move
-        if (tx > 0) {
-            // Target is to the right, strafe right
-            move(movePower, "right", Math.abs(strafeDistanceInches));
-        } else {
-            // Target is to the left, strafe left
-            move(movePower, "left", Math.abs(strafeDistanceInches));
-        }
-
-        // Restore manual control mode after movement
-        restoreManualControl();
-
-        // Verify we're centered
-        limelight.update();
-        if (limelight.isCentered()) {
-            opmode.telemetry.addData("Limelight Center", "Successfully centered by strafing!");
-            opmode.telemetry.update();
-            return true;
-        }
-
-        return true;
-    }
-
-    /**
-     * Center on Limelight target using both rotation and strafing (encoder-based)
-     * First rotates to face the target, then strafes to center position
-     * 
-     * @param limelight The Limelight instance
-     * @param rotationPower Power for rotation
-     * @param movePower Power for strafing
-     * @param distanceToTarget Estimated distance to target in inches
-     * @return true if successfully centered
-     */
-    public boolean centerOnLimelightTargetFull(org.firstinspires.ftc.teamcode.Game.Limelight limelight,
-                                                double rotationPower, double movePower, double distanceToTarget) {
-        // First, rotate to face the target
-        boolean rotated = centerOnLimelightTarget(limelight, rotationPower, 0);
-        
-        if (!rotated) {
-            restoreManualControl(); // Restore manual control even if rotation failed
-            return false;
-        }
-
-        opmode.sleep(200); // Small delay after rotation
-
-        // Then strafe to center position (this will restore manual control at the end)
-        return centerOnLimelightTargetByStrafing(limelight, movePower, distanceToTarget);
     }
